pyproject.toml
[tool.poetry]
name = "cmis_senario_games"
version = "0.1.0"
description = "Resilience cooperation game engine for virtual lifeline networks"
authors = ["cmis-team"]
readme = "README.md"
packages = [
  { include = "cmis_senario_games", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.0"
scipy = "^1.11.0"
networkx = "^3.2.0"
matplotlib = "^3.8.0"
pyyaml = "^6.0.0"
pandas = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ipykernel = "^6.29.0"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.poetry.scripts]
beta-eta-demo = "cmis_senario_games.cli.beta_eta_demo:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
